Add files via upload
Added ignition_timing_array to the main ttrs_quicfire source files. 

This file was requested by a partner and gives an array the size of the burn plot. Every cell is equal to -1 to start. The program reads ignite.dat to set an aerial or atv ignition. Every cell that gets ignited is updated to contain the time of ignition. Uses Bresenham's line algorithm to draw line ignitions and ignite appropriate cells. For example, a 5 x 5 plot with a line fire will look as follows. 
-1 -1 -1 -1 -1
-1  1 -1 -1 -1
-1  2 -1 -1 -1
-1  3 -1 -1 -1
-1 -1 -1 -1 -1
Where 1 is the start of the fire and 3 is the end. Each number is in seconds. 

This has been tested extensively for atv line fires. The line algorithm determines which direction the slope of the intended line moves and switches x and y values appropriately. Cell are then ignited if the slope of the original line does not stray too far from the slope calculated with that cell. This allows the line to be represented by cells with the same x value and different y values to give a much cleaner picture. Aerial fires also work and are much simpler and so did not require as extensive testing.